pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyper_lrrnn"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "numpy",
    "scikit-learn",
    "neurogym>=2.2.0",
    "lightning",
    "hydra-core",
    "wandb",
    "zuko>=1.4.1",
]

[tool.setuptools.packages.find]
include=["hyper_lrrnn"]